Move app stuff into own subpackage
