+++ fooddata_vegattributes/app/cli.py
