+++ fooddata_vegattributes/app/default_paths.py
