+++ setup.cfg
@@ -10,4 +10,4 @@ test = pytest~=7.1; mypy~=0.950; pytest-cov~=3.0
 
 [options.entry_points]
 console_scripts =
-  fooddata-vegattributes = fooddata_vegattributes.cli:main
+  fooddata-vegattributes = fooddata_vegattributes.app.cli:main
